An Event related performance improvement attempt, along with some code refactoring/simplifying

+++ server/src/com/openrsc/server/model/entity/player/Player.java
@@ -317,7 +317,7 @@ public final class Player extends Mob {
 	 */
 	private long sessionStart;
 
-	/**
+	/*
 	 * Restricts P2P stuff in F2P wilderness.
 	 */
 	/*public void unwieldMembersItems() {
@@ -489,9 +489,8 @@ public final class Player extends Mob {
 				if (entry.getValue() != p.getAppearanceID()) {
 					knownPlayersAppearanceIDs.put(p.getUsernameHash(), p.getAppearanceID());
 					return true;
-				} else {
-					return false;
 				}
+				return false;
 			}
 		}
 		knownPlayersAppearanceIDs.put(p.getUsernameHash(), p.getAppearanceID());
@@ -499,15 +498,9 @@ public final class Player extends Mob {
 	}
 
 	public boolean requiresAppearanceUpdateForPeek(final Player p) {
-		for (Entry<Long, Integer> entry : knownPlayersAppearanceIDs.entrySet()) {
-			if (entry.getKey() == p.getUsernameHash()) {
-				if (entry.getValue() != p.getAppearanceID()) {
-					return true;
-				} else {
-					return false;
-				}
-			}
-		}
+		for (Entry<Long, Integer> entry : knownPlayersAppearanceIDs.entrySet())
+			if (entry.getKey() == p.getUsernameHash())
+				return entry.getValue() != p.getAppearanceID();
 		return true;
 	}
 
@@ -564,7 +557,6 @@ public final class Player extends Mob {
 	public void addSkull(final long timeLeft) {
 		if (skullEvent == null) {
 			skullEvent = new DelayedEvent(getWorld(), this, timeLeft, "Player Add Skull") {
-
 				@Override
 				public void run() {
 					removeSkull();
@@ -715,15 +707,12 @@ public final class Player extends Mob {
 	}
 
 	public int combatStyleToIndex() {
-		if (getCombatStyle() == 1) {
+		if (getCombatStyle() == 1)
 			return 2;
-		}
-		if (getCombatStyle() == 2) {
+		if (getCombatStyle() == 2)
 			return 0;
-		}
-		if (getCombatStyle() == 3) {
+		if (getCombatStyle() == 3)
 			return 1;
-		}
 		return -1;
 	}
 
@@ -757,10 +746,8 @@ public final class Player extends Mob {
 					if (getOwner().canLogout() || (!(getOwner().inCombat() && getOwner().getDuel().isDuelActive())
 						&& System.currentTimeMillis() - startDestroy > 60000)) {
 						getOwner().unregister(true, reason);
-						running = false;
-					} else {
-						running = false;
 					}
+					running = false;
 				}
 			};
 			getWorld().getServer().getGameEventHandler().add(unregisterEvent);
@@ -1247,9 +1234,8 @@ public final class Player extends Mob {
 
 	public int getThrowingEquip() {
 		if (getWorld().getServer().getConfig().WANT_EQUIPMENT_TAB) {
-			Item item;
 			for (int i = 0; i < Equipment.slots; i++) {
-				item = getEquipment().get(i);
+				Item item = getEquipment().get(i);
 				if (item != null && DataConversions.inArray(Formulae.throwingIDs, item.getID())) {
 					return item.getID();
 				}
@@ -1337,8 +1323,6 @@ public final class Player extends Mob {
 			return getCache().getLong("skull_remaining");
 		if (!getCache().hasKey("skull_remaining"))
 			getSkullTime();
-		else
-			return 0;
 		return 0;
 	}
 
@@ -1348,9 +1332,8 @@ public final class Player extends Mob {
 	}
 
 	public long getChargeTime() {
-		if (isCharged()) {
+		if (isCharged())
 			return chargeEvent.timeTillNextRun();
-		}
 		return 0;
 	}
 
@@ -1359,8 +1342,6 @@ public final class Player extends Mob {
 			return getCache().getLong("charge_remaining");
 		if (!getCache().hasKey("charge_remaining"))
 			getChargeTime();
-		else
-			return 0;
 		return 0;
 	}
 
@@ -1385,7 +1366,6 @@ public final class Player extends Mob {
 	}
 
 	public String getUsername() {
-
 		return username;
 	}
 
@@ -1394,9 +1374,8 @@ public final class Player extends Mob {
 	}
 
 	public long getUsernameHash() {
-		if (getAttribute("fakeuser", null) != null) {
-			return DataConversions.usernameToHash((String) getAttribute("fakeuser", null));
-		}
+		if (getAttribute("fakeuser", null) != null)
+			return DataConversions.usernameToHash(getAttribute("fakeuser", null));
 		return usernameHash;
 	}
 
@@ -1474,9 +1453,8 @@ public final class Player extends Mob {
 
 	public void incQuestExp(final int i, final int amount) {
 		int appliedAmount = amount;
-		if (!isOneXp()) {
+		if (!isOneXp())
 			appliedAmount = (int) Math.round(getWorld().getServer().getConfig().SKILLING_EXP_RATE * amount);
-		}
 		if (isExperienceFrozen()) {
 			ActionSender.sendMessage(this, "You passed on " + appliedAmount/4 + " " +
 				getWorld().getServer().getConstants().getSkills().getSkill(i).getLongName() + " experience because your exp is frozen.");
@@ -1569,7 +1547,7 @@ public final class Player extends Mob {
 				//}
 				if (skill >= 4) {
 					fatigue += skillXP * 8;
-				} else if (skill >= 0 && skill <= 3) {
+				} else if (skill >= 0) {
 					fatigue += skillXP * 5;
 				}
 				if (fatigue > this.MAX_FATIGUE) {
@@ -1814,11 +1792,10 @@ public final class Player extends Mob {
 	}
 
 	public int getSkullType() {
-		int type = 0;
 		if (isSkulled()) {
-			type = 1;
+			return 1;
 		}
-		return type;
+		return 0;
 	}
 
 	public boolean isSleeping() {
@@ -1851,10 +1828,9 @@ public final class Player extends Mob {
 		ActionSender.sendSound(this, "death");
 		ActionSender.sendDied(this);
 
-		if (getAttribute("projectile", null) != null) {
-			ProjectileEvent projectileEvent = getAttribute("projectile");
+		ProjectileEvent projectileEvent = getAttribute("projectile");
+		if (projectileEvent != null)
 			projectileEvent.setCanceled(true);
-		}
 		getSettings().getAttackedBy().clear();
 
 		getCache().store("last_death", System.currentTimeMillis());
@@ -2198,7 +2174,7 @@ public final class Player extends Mob {
 
 		save();
 
-		/** IP Tracking in wilderness removal */
+		/* IP Tracking in wilderness removal */
 		/*if(player.getLocation().inWilderness())
 		{
 			wildernessIPTracker.remove(player.getCurrentIP());
@@ -2249,13 +2225,10 @@ public final class Player extends Mob {
 
 	public void setBatchEvent(final BatchEvent batchEvent) {
 		if (batchEvent != null && batchEvent.getOwner() != null) {
-			Player player = (Player) batchEvent.getOwner();
-			player.checkAndInterruptBatchEvent();
-			this.batchEvent = batchEvent;
+			batchEvent.getOwner().checkAndInterruptBatchEvent();
 			getWorld().getServer().getGameEventHandler().add(batchEvent);
-		} else {
-			this.batchEvent = null;
 		}
+		this.batchEvent = batchEvent;
 	}
 
 	public void setCastTimer(final long timer) {
@@ -2399,14 +2372,11 @@ public final class Player extends Mob {
 
 	@Override
 	public void setLocation(final Point p, final boolean teleported) {
-		if (!teleported) {
-			if (getSkullType() == 2)
-				getUpdateFlags().setAppearanceChanged(true);
-			else if (getSkullType() == 0)
-				getUpdateFlags().setAppearanceChanged(true);
-		} else {
-			setTeleporting(true);
+		if (teleported || getSkullType() == 2 || getSkullType() == 0) {
+			// Inappropriate place for this to be getting set at for skulls, to me.
 			getUpdateFlags().setAppearanceChanged(true);
+			if (teleported)
+				setTeleporting(true);
 		}
 
 		super.setLocation(p, teleported);
@@ -2858,15 +2828,15 @@ public final class Player extends Mob {
 				return 0x03FFFFFF;
 
 			return 0;
-		} else {
-			if (isAdmin())
-				return 0x02FFFFFF;
-
-			if (isMod())
-				return 0x03FFFFFF;
-
-			return 0;
 		}
+
+		if (isAdmin())
+			return 0x02FFFFFF;
+
+		if (isMod())
+			return 0x03FFFFFF;
+
+		return 0;
 	}
 
 	public Trade getTrade() {
@@ -2940,16 +2910,10 @@ public final class Player extends Mob {
 	}
 
 	public void addNpcKill(final Npc n, final boolean sendUpdate) {
-		int kills = 1;
-		if (getKillCache().containsKey(n.getID())) {
-			kills = getKillCache().get(n.getID()) + 1;
-
-		}
+		int kills = getKillCache().containsKey(n.getID()) ? getKillCache().get(n.getID()) + 1 : 1;
 		getKillCache().put(n.getID(), kills);
 		setKillCacheUpdated(true);
-		if (sendUpdate) {
-			message("Your " + n.getDef().getName() + " kill count is: @red@" + kills + "@whi@.");
-		}
+		if (sendUpdate) message("Your " + n.getDef().getName() + " kill count is: @red@" + kills + "@whi@.");
 	}
 
 	public boolean hasHigherRankThan(final Entity observer) {
